Add Asset Service , fix bugs

+++ src/main/java/com/nt/service/OrderServiceImpl.java
@@ -2,10 +2,7 @@ package com.nt.service;
 
 import com.nt.domain.OrderStatus;
 import com.nt.domain.OrderType;
-import com.nt.model.Coin;
-import com.nt.model.Order;
-import com.nt.model.OrderItem;
-import com.nt.model.User;
+import com.nt.model.*;
 import com.nt.repository.OrderItemRepository;
 import com.nt.repository.OrderRepository;
 import jakarta.transaction.Transactional;
@@ -27,6 +24,9 @@ public class OrderServiceImpl implements OrderService {
 
     @Autowired
     private OrderItemRepository orderItemRepository ;
+
+    @Autowired
+    private AssetService assetService ;
 
     @Override
     public Order createOrder(User user, OrderItem orderItem, OrderType orderType) {
@@ -77,6 +77,15 @@ public class OrderServiceImpl implements OrderService {
         Order savedOrder = orderRepository.save(order);
 
         // create asset
+        Asset oldAsset = assetService.findAssetByUserIdAndCoinId(order.getUser().getId(),
+                                                                 order.getOrderItem().getCoin().getId()
+                                                                 );
+        if (oldAsset == null) {
+            assetService.createAsset(user , orderItem.getCoin() , orderItem.getQuantity() ) ;
+        }
+        else{
+            assetService.updateAsset(oldAsset.getId() , quantity) ;
+        }
 
         return savedOrder ;
     }
@@ -87,25 +96,29 @@ public class OrderServiceImpl implements OrderService {
             throw new RuntimeException("Quantity must be greater than 0");
         }
         double sellPrice = coin.getCurrentPrice() ;
+        Asset assetToSell = assetService.findAssetByUserIdAndCoinId(user.getId() , coin.getId()) ;
+        double buyPrice = assetToSell.getBuyPrice();
+        if(assetToSell != null) {
+            OrderItem orderItem = createOrderItem(coin, quantity, buyPrice, sellPrice);
+            Order order = createOrder(user, orderItem, OrderType.SELL);
 
-        double buyPrice = assetToSell.getPrice() ;
+            orderItem.setOrder(order);
 
-        OrderItem orderItem = createOrderItem(coin, quantity, buyPrice ,sellPrice);
-        Order order = createOrder(user, orderItem, OrderType.SELL);
-        orderItem.setOrder(order);
+            if (assetToSell.getQuantity() >= quantity) {
+                order.setStatus(OrderStatus.SUCCESSFUL);
+                order.setOrderType(OrderType.SELL);
+                Order savedOrder = orderRepository.save(order);
+                walletService.payOrderPayment(order, user);
 
-        if(assetToSell.getQuantity() >= quantity){
-            order.setStatus(OrderStatus.SUCCESSFUL);
-            order.setOrderType(OrderType.SELL);
-            Order savedOrder = orderRepository.save(order);
-            walletService.payOrderPayment(order, user);
-            Asset updatedAsset = assetService.updateAsset(assetToSell.getId(), -quantity);
-            if(updatedAsset.getQuantity()*coin.getCurrentPrice() <= 1){
-                assetService.deleteAsset(updatedAsset.getId());
+                Asset updatedAsset = assetService.updateAsset(assetToSell.getId(), -quantity);
+                if (updatedAsset.getQuantity() * coin.getCurrentPrice() <= 1) {
+                    assetService.deleteAsset(updatedAsset.getId());
+                }
+                return savedOrder;
             }
-            return savedOrder;
+            throw new RuntimeException("You don't have enough quantity to sell");
         }
-        throw new RuntimeException("You don't have enough quantity to sell");
+        throw new RuntimeException("You don't have this asset to sell");
     }
 
     @Override
